:recycle: refactor(GA): modify google analystics event handling
