+++ next-env.d.ts
@@ -1,17 +1,6 @@
 /// <reference types="next" />
 /// <reference types="next/types/global" />
+/// <reference types="next/image-types/global" />
 
-interface Window {
-  // For pageview
-  gtag(type: 'config', googleAnalyticsId: string, { page_path: string })
-  // For event
-  gtag(
-    type: 'event',
-    eventAction: string,
-    fieldObject: {
-      event_label: string
-      event_category: string
-      value?: number
-    }
-  )
-}
+// NOTE: This file should not be edited
+// see https://nextjs.org/docs/basic-features/typescript for more information.
