+++ pages/_app.tsx
@@ -1,23 +1,16 @@
 import { AppProps } from 'next/app'
-import { useEffect } from 'react'
-import { useRouter } from 'next/router'
-import * as gtag from '../lib/gtag'
+import { GoogleAnalytics, usePageView } from '../lib/gtag'
 import '../styles/globals.css'
 
 const App = ({ Component, pageProps }: AppProps) => {
-  const router = useRouter()
-  useEffect(() => {
-    const handleRouteChange = (path: string) => {
-      gtag.pageview(path)
-    }
+  usePageView()
 
-    router.events.on('routeChangeComplete', handleRouteChange)
-    return () => {
-      router.events.off('routeChangeComplete', handleRouteChange)
-    }
-  }, [router.events])
-
-  return <Component {...pageProps} />
+  return (
+    <>
+      <GoogleAnalytics />
+      <Component {...pageProps} />
+    </>
+  )
 }
 
 export default App
